requirements.txt
airtest==1.1.1
allure-pytest==2.8.11
allure-python-commons==2.8.11
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
comtypes==1.1.7
cryptography==2.8
Deprecated==1.2.7
Django==2.2.5
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
facebook-wda==0.2.1
Flask==1.1.1
gitdb2==2.0.6
GitPython==3.0.5
hrpc==1.0.8
idna==2.7
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
MarkupSafe==1.1.1
more-itertools==7.2.0
mss==4.0.3
nose==1.3.7
numpy==1.17.4
opencv-contrib-python==3.4.2.17
packaging==19.2
paramiko==2.6.0
Pillow==6.2.1
pluggy==0.13.1
poco==0.96.8
pocoui==1.0.78
pushbullet.py==0.11.0
py==1.8.0
pyaml==19.4.1
pycparser==2.19
PyGithub==1.44.1
PyJWT==1.7.1
pylinktester==0.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.5
pypiwin32==223
pytest==5.3.0
pytest-html==2.0.1
pytest-metadata==1.8.0
python-dateutil==2.8.1
python-gitlab==1.13.0
python-magic==0.4.15
pytz==2019.2
pywin32==225
pywinauto==0.6.3
PyYAML==5.1.2
requests==2.20.1
selenium==3.141.0
six==1.13.0
smmap2==2.0.5
sqlparse==0.3.0
svn==0.3.46
texttable==0.9.1
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.15.6
wrapt==1.11.2
xlrd==1.2.0
zipp==0.6.0
